[Fix] #52 - divide userInfoModel DTO
